feat: add dedicated index and search service modules
